cloud-function/requirements.txt
google-api-python-client==1.7.11
google-cloud-container==0.3.0
kubernetes==10.0.1
Flask==1.1.1
